test: move Python e2e scenarios to matrix
Core (make test-e2e) was running language-specific Python cases
under TestCore_*, including pack/s2i and remote cells, so the
"E2E - Core, Metadata, and Remote" job paid ~13m for work that
belongs on E2E - Runtimes (python).

Rename those six tests to TestMatrix_Python_* and gate them with
requireMatrixRuntime(t, "python") so they ride make test-e2e-matrix
(and -run TestMatrix_Python for local subsets). Makefile and
workflow filters need no change.

Keep CredentialsDockerPusher and Remote in the core job for now.

Functions #58
