+++ e2e/README.md
@@ -21,14 +21,16 @@ acceptance criteria for pull requests.
 Tests themselves are separated into categories:  Core, Metadata,
 Remote, Podman, Matrix, and Config CI.
 
-Core tests include checking the basic CRUDL operations; Create, Read, Update,
-Delete and List.  Creation is implemented as `func init`, running the function
-locally with `func run`, and running the cluster with `func deploy`. Reading is
+Core tests cover the go-centric CRUDL path: Create, Read, Update, Delete and
+List.  Creation is implemented as `func init`, running the function locally
+with `func run`, and running on the cluster with `func deploy`. Reading is
 implemented as `func describe`.  Updating, which ensures that an updated
 function replaces the old, is implemented as `func deploy`.  Finally,
 `func list` implements a standard listing operation.  Tests also confirm
 the core features function when referring to a remote repository, using
-templates (repositories) and other core variants.
+templates (repositories) and other core variants.  Language-specific depth
+(Python signatures, userdeps, legacy parliament, etc.) lives under Matrix,
+not Core.
 
 Metadata tests ensure that manipulation of a Function's metadata is correctly
 carried to the final Function.  Metadata includes environment variables,
@@ -40,12 +42,12 @@ via in-cluster builds, etc.
 Podman tests ensure that the Podman container engine is also supported. Note
 that these tests require that 'podman' and 'ssh' are available in your path.
 
-Matrix tests is a larger set which checks operations which differ in
-implementation between language runtimes.  The primary operations which
-differ and must be checked for each runtime are creation and running locally.
-Therefore, the runtime tests execute for each language, for each template, for
-each builder.  As a side-effect of the test implementation, "func invoke" is
-also tested.
+Matrix tests cover operations that differ by language runtime.  The primary
+permutations (each runtime × template × builder) check create, local run,
+deploy, and remote deploy.  In addition, language-specific scenarios are
+named `TestMatrix_<Runtime>_*` (for example `TestMatrix_Python_*`) so they
+run on the corresponding runtime matrix job rather than in Core, and can be
+selected with a prefix filter.
 
 Config CI tests verify that GitHub Actions workflows generated by
 `func config ci` deploy a function successfully. They use nektos/act
@@ -225,6 +227,7 @@ then `go test -tags e2e -run TestCore ./e2e`. Subsets include:
 - TestMetadata
 - TestRemote
 - TestMatrix
+- TestMatrix_Python  (Python-specific matrix scenarios; requires FUNC_E2E_MATRIX=true)
 - TestConfigCI
 
 ## Cleanup
