+++ e2e/e2e_core_test.go
@@ -374,120 +374,6 @@ func Handle(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-// TestCore_StaticSignature_Python ensures backward compatibility with the
-// static (non-instanced) Python function signature. Python functions can
-// export either a `new()` constructor (instanced) or a plain `handle()`
-// function (static). The scaffolding imports `new` first and falls back
-// to `handle` on ImportError.
-func TestCore_StaticSignature_Python(t *testing.T) {
-	name := "func-e2e-test-core-static-py"
-	root := fromCleanEnv(t, name)
-
-	if err := newCmd(t, "init", "-l=python", "-t=http").Run(); err != nil {
-		t.Fatal(err)
-	}
-
-	staticPy := `async def handle(scope, receive, send):
-    await send({
-        'type': 'http.response.start',
-        'status': 200,
-        'headers': [[b'content-type', b'text/plain']],
-    })
-    await send({
-        'type': 'http.response.body',
-        'body': b'OK',
-    })
-`
-	if err := os.WriteFile(filepath.Join(root, "function", "func.py"), []byte(staticPy), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(root, "function", "__init__.py"), []byte("from .func import handle\n"), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	if err := newCmd(t, "deploy", "--builder", "host").Run(); err != nil {
-		t.Fatal(err)
-	}
-	defer func() {
-		clean(t, name, Namespace)
-	}()
-
-	if !waitFor(t, ksvcUrl(name)) {
-		t.Fatal("static Python function did not deploy correctly")
-	}
-}
-
-// LEGACY PYTHON: old parliament functions (func.py def main(context) + parliament
-// Procfile) must build and serve via pack and s2i, built locally and in-cluster.
-func TestCore_LegacyParliamentSignature_Python(t *testing.T) {
-	// The fixtures in testdata/legacy-parliament were generated with func
-	// v1.17.0 (the last parliament-era release):
-	//
-	//	func create -l python -t http parliament-http
-	//	func create -l python -t cloudevents parliament-cloudevents
-	//
-	// The cells pair builders, flavors and localities so that each builder
-	// builds both locally and remotely (in-cluster, where scaffolding runs
-	// inside the Tekton pipeline) and each flavor passes through both builders.
-	for _, tc := range []struct {
-		builder  string
-		template string
-		remote   bool
-	}{
-		{"pack", "http", false},
-		{"s2i", "cloudevents", false},
-		{"pack", "cloudevents", true},
-		{"s2i", "http", true},
-	} {
-		cell := tc.builder
-		if tc.remote {
-			cell += "-remote"
-		}
-		t.Run(cell, func(t *testing.T) {
-			name := "func-e2e-parliament-" + cell
-			root := fromCleanEnv(t, name)
-
-			// Copy in the v1.17.0-generated fixture, renaming it for this run;
-			// the rest of the era func.yaml stays as generated.
-			fixture := "parliament-" + tc.template
-			if err := os.CopyFS(root, os.DirFS(filepath.Join(Testdata, "legacy-parliament", fixture))); err != nil {
-				t.Fatal(err)
-			}
-			yamlPath := filepath.Join(root, "func.yaml")
-			y, err := os.ReadFile(yamlPath)
-			if err != nil {
-				t.Fatal(err)
-			}
-			y = bytes.Replace(y, []byte("name: "+fixture), []byte("name: "+name), 1)
-			if err := os.WriteFile(yamlPath, y, 0644); err != nil {
-				t.Fatal(err)
-			}
-
-			args := []string{"deploy", "--builder", tc.builder}
-			if tc.remote {
-				args = append(args, "--remote")
-			}
-			if err := newCmd(t, args...).Run(); err != nil {
-				t.Fatal(err)
-			}
-			defer func() {
-				clean(t, name, Namespace)
-			}()
-
-			if tc.template == "cloudevents" {
-				if !waitFor(t, ksvcUrl(name), withTemplate("cloudevents")) {
-					t.Fatalf("legacy parliament cloudevents function did not deploy correctly with %s builder", tc.builder)
-				}
-			} else {
-				// The parliament http template echoes the query as JSON.
-				if !waitFor(t, ksvcUrl(name)+"?foo=bar", withContentMatch(`{"foo": "bar"}`)) {
-					t.Fatalf("legacy parliament http function did not deploy correctly with %s builder", tc.builder)
-				}
-			}
-		})
-	}
-}
-
 // TestCore_Delete ensures that a function registered as deleted when deleted.
 // Also tests list as a side-effect.
 //
